+++ src/_types/FetchHttpHeaders.d.ts
@@ -1,3 +1,3 @@
 export interface FetchHttpHeaders extends Record<string, string> {
-	'Content-Type':string;
+	'Content-Type':string
 }