+++ src/_types/FetchHttpOpts.d.ts
@@ -1,5 +1,5 @@
 import type { FetchHttpHeaders } from './FetchHttpHeaders'
 export interface FetchHttpOpts {
-	url?:string;
-	headers?:FetchHttpHeaders;
+	url?:string
+	headers?:FetchHttpHeaders
 }