+++ src/graphql_url_/index.d.ts
@@ -1,0 +1,5 @@
+export declare function graphql_url_(host?:string, port?:number):string
+export {
+	graphql_url_ as _graphql_url,
+	graphql_url_ as _url__graphql,
+}